Add files via upload
This is my website created in a school project at UIB. It is a website for tourists in Bergen with information about the most popular attractions. You can also find location and availability information about public toilets. There  is also a map showing playgrounds located in Bergen. The page for favorite playground is not quite done.